requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
greenlet==3.1.1
h11==0.14.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.63.2
pika==1.3.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
requests==2.32